pyproject.toml
[tool.poetry]
name = "python-test-luizalabs"
version = "0.1.0"
description = ""
authors = ["Elson Silva <elsonjuniog@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.122.0"}
asyncpg = "^0.31.0"
sqlalchemy = "^2.0.44"
pydantic = "^2.12.4"
python-dotenv = "^1.2.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"
alembic = "^1.17.2"
psycopg2-binary = "^2.9.11"
pydantic-settings = "^2.12.0"
email-validator = "^2.3.0"
bcrypt = "4.2.0"
requests = "^2.32.5"
cryptography = "^46.0.3"
redis = "^7.1.0"
uvicorn = "^0.38.0"
python-json-logger = "^4.0.0"


[tool.poetry.group.dev.dependencies]
alembic = "^1.17.2"
pytest = "^9.0.1"
pytest-asyncio = "^1.3.0"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
